Further refactoring of client and server code breaks.
