+++ src/main/java/io/github/tofodroid/mods/mimi/client/gui/GuiMidiInputConfig.java
@@ -4,8 +4,8 @@ import javax.sound.midi.MidiDevice.Info;
 
 import com.mojang.blaze3d.matrix.MatrixStack;
 
+import io.github.tofodroid.mods.mimi.client.midi.MidiInputManager;
 import io.github.tofodroid.mods.mimi.common.MIMIMod;
-import io.github.tofodroid.mods.mimi.common.midi.MidiInputManager;
 import net.minecraft.client.Minecraft;
 import net.minecraft.entity.player.PlayerEntity;
 import net.minecraft.util.math.vector.Vector2f;
@@ -21,7 +21,7 @@ public class GuiMidiInputConfig extends BaseGui {
 
     public GuiMidiInputConfig(PlayerEntity player) {
         super(300, 158, 300, "textures/gui/gui_midi_config.png",  "item.MIMIMod.gui_midi_input_config");
-        this.midiInputManager = MIMIMod.proxy.getMidiInput();
+        this.midiInputManager = (MidiInputManager)MIMIMod.proxy.getMidiInput();
     }
 
     @Override
