+++ src/main/java/io/github/tofodroid/mods/mimi/server/ServerProxy.java
@@ -1,8 +1,8 @@
 package io.github.tofodroid.mods.mimi.server;
 
+import io.github.tofodroid.mods.mimi.client.midi.MidiInputManager;
 import io.github.tofodroid.mods.mimi.client.midi.MidiSynthManager;
 import io.github.tofodroid.mods.mimi.common.Proxy;
-import io.github.tofodroid.mods.mimi.common.midi.MidiInputManager;
 import net.minecraftforge.fml.event.lifecycle.FMLCommonSetupEvent;
 
 public class ServerProxy implements Proxy {
